olist_dwh/requirements.txt
psycopg2-binary>=2.9
python-dateutil>=2.8
